refactor: separate 'notification' and 'notice' and change tooltip position
