+++ apps/web/src/common/modules/navigations/top-bar/modules/top-bar-toolset/TopBarToolset.vue
@@ -11,7 +11,9 @@ import { useAppContextStore } from '@/store/app-context/app-context-store';
 import TopBarAdminToggleButton from '@/common/modules/navigations/top-bar/modules/top-bar-toolset/modules/top-bar-admin-toggle-button/TopBarAdminToggleButton.vue';
 import TopBarFavorite
     from '@/common/modules/navigations/top-bar/modules/top-bar-toolset/modules/top-bar-favorite/TopBarFavorite.vue';
-import TopBarNoti from '@/common/modules/navigations/top-bar/modules/top-bar-toolset/modules/top-bar-noti/TopBarNoti.vue';
+import TopBarNotice
+    from '@/common/modules/navigations/top-bar/modules/top-bar-toolset/modules/top-bar-notice/TopBarNotice.vue';
+import TopBarNotifications from '@/common/modules/navigations/top-bar/modules/top-bar-toolset/modules/top-bar-notifications/TopBarNotifications.vue';
 import TopBarProfile from '@/common/modules/navigations/top-bar/modules/top-bar-toolset/modules/top-bar-profile/TopBarProfile.vue';
 
 const props = withDefaults(defineProps<{
@@ -29,9 +31,6 @@ const state = reactive({
     isAdminMode: computed(() => appContextStore.getters.isAdminMode),
     isGrantLoading: computed(() => appContextStore.getters.globalGrantLoading),
     tooltipTexts: computed<Record<string, string>>(() => ({
-        recentFavorite: i18n.t('COMMON.GNB.TOOLTIP.RECENT_FAVORITE') as string,
-        notifications: i18n.t('COMMON.GNB.TOOLTIP.NOTIFICATIONS') as string,
-        profile: i18n.t('COMMON.GNB.TOOLTIP.PROFILE') as string,
         adminToggle: (state.isAdminMode ? i18n.t('COMMON.GNB.TOOLTIP.EXIT_ADMIN_MODE') : i18n.t('COMMON.GNB.TOOLTIP.ENABLE_ADMIN_MODE')) as string,
     })),
 });
@@ -51,35 +50,26 @@ const updateOpenedMenu = (menu: string, visible: boolean) => {
 
 <template>
     <div class="top-bar-toolset">
-        <!-- TODO: top-bar-search-clone-->
-        <p-tooltip :contents="state.tooltipTexts.recentFavorite"
-                   position="bottom"
-        >
-            <top-bar-favorite v-if="!state.isAdminMode && !state.isGrantLoading"
-                              :visible="props.openedMenu === 'recentFavorite'"
-                              @update:visible="updateOpenedMenu('recentFavorite', $event)"
-            />
-        </p-tooltip>
-        <p-tooltip :contents="state.tooltipTexts.notifications"
-                   position="bottom"
-        >
-            <top-bar-noti v-if="!state.isAdminMode && !state.isGrantLoading"
-                          :visible="props.openedMenu === 'notifications'"
-                          @update:visible="updateOpenedMenu('notifications', $event)"
-            />
-        </p-tooltip>
+        <top-bar-favorite v-if="!state.isAdminMode && !state.isGrantLoading"
+                          :visible="props.openedMenu === 'favorite'"
+                          @update:visible="updateOpenedMenu('favorite', $event)"
+        />
+        <top-bar-notifications v-if="!state.isAdminMode && !state.isGrantLoading"
+                               :visible="props.openedMenu === 'notifications'"
+                               @update:visible="updateOpenedMenu('notifications', $event)"
+        />
+        <top-bar-notice v-if="!state.isAdminMode && !state.isGrantLoading"
+                        :visible="props.openedMenu === 'notice'"
+                        @update:visible="updateOpenedMenu('notice', $event)"
+        />
         <p-tooltip :contents="state.tooltipTexts.adminToggle"
                    position="bottom"
         >
             <top-bar-admin-toggle-button v-if="state.isDomainAdmin" />
         </p-tooltip>
-        <p-tooltip :contents="state.tooltipTexts.profile"
-                   position="bottom"
-        >
-            <top-bar-profile :visible="props.openedMenu === 'profile'"
-                             @update:visible="updateOpenedMenu('profile', $event)"
-            />
-        </p-tooltip>
+        <top-bar-profile :visible="props.openedMenu === 'profile'"
+                         @update:visible="updateOpenedMenu('profile', $event)"
+        />
     </div>
 </template>
